+++ app/Api/V1_0/Products/Controllers/ProductController.php
@@ -8,6 +8,8 @@ use ERP\Api\V1_0\Support\BaseController;
 use ERP\Api\V1_0\Products\Processors\ProductProcessor;
 use ERP\Core\Products\Persistables\ProductPersistable;
 use ERP\Core\Support\Service\ContainerInterface;
+use ERP\Exceptions\ExceptionMessage;
+use ERP\Model\Products\ProductModel;
 /**
  * @author Reema Patel<reema.p@siliconbrain.in>
  */
@@ -15,13 +17,13 @@ class ProductController extends BaseController implements ContainerInterface
 {
 	/**
      * @var productService
-     * @var Processor
+     * @var processor
      * @var name
      * @var request
      * @var productPersistable
      */
 	private $productService;
-	private $Processor;
+	private $processor;
 	private $request;
 	private $productPersistable;	
 	
@@ -50,10 +52,10 @@ class ProductController extends BaseController implements ContainerInterface
 		// insert
 		if($requestMethod == 'POST')
 		{
-			$Processor = new ProductProcessor();
+			$processor = new ProductProcessor();
 			$productPersistable = new ProductPersistable();		
 			$productService= new ProductService();			
-			$productPersistable = $Processor->createPersistable($this->request);
+			$productPersistable = $processor->createPersistable($this->request);
 			if($productPersistable[0][0]=='[')
 			{
 				return $productPersistable;
@@ -132,22 +134,31 @@ class ProductController extends BaseController implements ContainerInterface
 	public function update(Request $request,$productId)
     {    
 		$this->request = $request;
-		$Processor = new ProductProcessor();
+		$processor = new ProductProcessor();
 		$productPersistable = new ProductPersistable();		
 		$productService= new ProductService();			
-		$productPersistable = $Processor->createPersistableChange($this->request,$productId);
-		if($productPersistable=="204: No Content Found For Update")
+		$productModel = new ProductModel();
+		$result = $productModel->getData($productId);
+		
+		//get exception message
+		$exception = new ExceptionMessage();
+		$fileSizeArray = $exception->messageArrays();
+		if(strcmp($result,$fileSizeArray['404'])==0)
 		{
-			return $productPersistable;
-		}
-		else if(is_array($productPersistable))
-		{
-			$status = $productService->update($productPersistable);
-			return $status;
+			return $result;
 		}
 		else
 		{
-			return $productPersistable;
+			$productPersistable = $processor->createPersistableChange($this->request,$productId);
+			if(is_array($productPersistable))
+			{
+				$status = $productService->update($productPersistable);
+				return $status;
+			}
+			else
+			{
+				return $productPersistable;
+			}
 		}
 	}
 	
@@ -158,12 +169,25 @@ class ProductController extends BaseController implements ContainerInterface
     public function Destroy(Request $request,$productId)
     {
         $this->request = $request;
-		$Processor = new ProductProcessor();
+		$processor = new ProductProcessor();
 		$productPersistable = new ProductPersistable();		
-		$productService= new ProductService();			
-		$productPersistable = $Processor->createPersistableChange($this->request,$productId);
-		$productService->create($productPersistable);
-		$status = $productService->delete($productPersistable);
-		return $status;
+		$productService= new ProductService();		
+		$productModel = new ProductModel();
+		$result = $productModel->getData($productId);
+		
+		//get exception message
+		$exception = new ExceptionMessage();
+		$fileSizeArray = $exception->messageArrays();
+		if(strcmp($result,$fileSizeArray['404'])==0)
+		{
+			return $result;
+		}
+		else
+		{		
+			$productPersistable = $processor->createPersistableChange($this->request,$productId);
+			$productService->create($productPersistable);
+			$status = $productService->delete($productPersistable);
+			return $status;
+		}
     }
 }
